de-CTKS map tests
